Start refactor of package to use Typescript
