+++ .gitignore
@@ -1,5 +1,5 @@
 .idea
 images
 node_modules
+coverage
 .DS_Store
-.nyc_output
