+++ .tool-versions
@@ -1,0 +1,1 @@
+nodejs 16.15.1
