--- .travis.yml
@@ -1,5 +1,0 @@
-language: node_js
-node_js:
-  - "10"
-  - "12"
-  - "14"
