+++ package.json
@@ -1,6 +1,9 @@
 {
   "name": "paparazzi-cli",
   "version": "0.4.0",
+  "engines": {
+    "node": ">=16.0.0"
+  },
   "description": "CLI tool to snap all the angles of your website",
   "main": "index.js",
   "author": "Edward Dobson",
@@ -11,20 +14,26 @@
     "screenshot",
     "cli"
   ],
+  "bin": {
+    "paparazzi": "./src/bin/paparazzi.js"
+  },
   "scripts": {
-    "test": "nyc -x **/*.spec.js mocha **/*.spec.js"
+    "test": "jest --collectCoverage"
   },
   "dependencies": {
-    "puppeteer": "^1.14.0",
-    "yargs": "^13.2.2"
-  },
-  "bin": {
-    "paparazzi": "./index.js"
+    "puppeteer": "^14.4.1",
+    "yargs": "^17.5.1"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "mocha": "^7.1.1",
-    "nyc": "^15.0.0",
-    "sinon": "^7.3.1"
+    "@types/jest": "^28.1.2",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "eslint": "^8.18.0",
+    "eslint-config-eslint": "^7.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "jest": "^28.1.1",
+    "jest-mock-extended": "^2.0.6",
+    "ts-jest": "^28.0.5",
+    "typescript": "^4.7.4"
   }
 }
