+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+      "target": "es6",
+      "module": "commonjs",
+      "lib": ["es6", "es2018", "dom"],
+      "declaration": true,
+      "outDir": "dist/lib",
+      "rootDir": "src",
+      "strict": true,
+      "types": ["jest", "node"],
+      "esModuleInterop": true,
+      "resolveJsonModule": true
+    }
+}
